1440: Move OBAddressType2Code to common pkg
OBAddressType2Code is shared across account, payment and vrp schemas.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1440
