--- secure-api-gateway-ob-uk-common-obie-datamodel/src/main/java/uk/org/openbanking/datamodel/v4/account/OBAddressType2Code.java
@@ -1,72 +1,0 @@
-/*
- * Copyright © 2020-2024 ForgeRock AS (obst@forgerock.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package uk.org.openbanking.datamodel.v4.account;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonValue;
-
-import jakarta.annotation.Generated;
-
-/**
- * Identifies the nature of the postal address. For a full set of codes see `OBAddressType2Code` [here](https://github.com/OpenBankingUK/External_Internal_CodeSets).
- */
-
-@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
-public enum OBAddressType2Code {
-
-    BIZZ("BIZZ"),
-
-    DLVY("DLVY"),
-
-    MLTO("MLTO"),
-
-    PBOX("PBOX"),
-
-    ADDR("ADDR"),
-
-    HOME("HOME"),
-
-    CORR("CORR"),
-
-    STAT("STAT");
-
-    private String value;
-
-    OBAddressType2Code(String value) {
-        this.value = value;
-    }
-
-    @JsonValue
-    public String getValue() {
-        return value;
-    }
-
-    @Override
-    public String toString() {
-        return String.valueOf(value);
-    }
-
-    @JsonCreator
-    public static OBAddressType2Code fromValue(String value) {
-        for (OBAddressType2Code b : OBAddressType2Code.values()) {
-            if (b.value.equals(value)) {
-                return b;
-            }
-        }
-        throw new IllegalArgumentException("Unexpected value '" + value + "'");
-    }
-}
-
