+++ secure-api-gateway-ob-uk-common-obie-datamodel/src/main/java/uk/org/openbanking/datamodel/v4/common/OBPostalAddress7.java
@@ -26,7 +26,6 @@ import jakarta.annotation.Generated;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
-import uk.org.openbanking.datamodel.v4.payment.OBAddressType2Code;
 
 /**
  * Information that locates and identifies a specific address, as defined by postal services.
